+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 4.0.0
+
+Package now depends on `aldeed:simple-schema` instead of NPM package.
+Async fixes for Meteor 3, minimum required Meteor version is 2.8.1
+
 ## 3.1.1
 
 Added compatibility option with `aldeed:collection2@4.0.4`